added IsValid() stub
We want to enforce all subclasses to have a public static IsValid() method,
but it's not possible to define a method both 'abstract' and 'static'.
So we do the next best thing:
Have the method in the base class and have it throw an ErrorException,
forcing the subclass author to override it with their own
implementation.