use PHP7 language features
This is the end of the v1.x branch
because this commit introduces BC-breaking changes.

Most importantly, the `isValid()` method
is no longer a stub in the `AbstractValue` base class;
it's now an `abstract static` method,
implicitly forcing users of the class to implement that method
without the need for a `NotImplementedException`.

Additionally, the AbstractValue methods
now have return types.